maximum spread matching method is in changed app from N processes to multiThreaded for local test

+++ memeMatcher.py
@@ -13,13 +13,24 @@ import unittest
 import tempfile
 import json
 import random
+import md5
+
 import requests
 import pyechonest.track
+
 import pygn
 
+from itertools import izip_longest
 from bs4 import BeautifulSoup
 
 GRACENOTE_KEY = os.environ['GRACENOTE_KEY']
+
+
+
+def grouper(n, iterable, fillvalue=None):
+    "grouper(3, 'ABCDEFG', 'x') --> ABC DEF Gxx"
+    args = [iter(iterable)] * n
+    return izip_longest(fillvalue=fillvalue, *args)
 
 def base_path():
     return os.path.dirname(os.path.realpath(__file__))
@@ -75,7 +86,8 @@ class memeMatcher:
     #mild cleaning
     artist = self.artist
     title = self.title.split('/')[0]
-    title = " ".join(title.split(" feat")[:-1])
+    if " feat" in title:
+      title = " ".join(title.split(" feat"))
     self.lyrics = rep_genius_parser(artist, title)
 
   def _fetch_cover_art(self):
@@ -83,36 +95,45 @@ class memeMatcher:
     userID = pygn.register(clientID)
     metadata = pygn.search(clientID=clientID, userID=userID, artist=self.artist, track=self.title.split('/')[0])
     try:
-      r = requests.get(metadata['album_art_url'])
-      artist_pic = tempfile.mkstemp(dir=os.path.join(base_path(), 'tmp'))[1]
-      with open(artist_pic, 'wb') as wh:
-        wh.write(r.content)
+      artist_pic = os.path.join(base_path(), 'tmp', md5.md5(metadata['album_art_url']).hexdigest())
+      if not os.path.exists(artist_pic):
+        r = requests.get(metadata['album_art_url'])
+        with open(artist_pic, 'wb') as wh:
+          wh.write(r.content) 
       self.album_art = '/audio_files/'+os.path.split(artist_pic)[1]
     except KeyError:
       print 'for', self.artist, ':', self.track, "gracenote doesn't have any album art"
       self.album_art = None
 
-  def select_and_align_memes(self):
+  def select_and_align_memes(self, method='maximum_spread'):
     with open(os.path.join(base_path(),'images.json')) as rh:
       memes = json.loads(rh.read())['memes']
-    if self.album_art:
-      self.timings = [{"image_url":self.album_art,
-                        "transition_after": int(self.track.sections[0]['duration']*1000),
-                        "top_text": '',
-                        "bottom_text": '',
-                      }] 
-      start_section = 1
-    else:
-      self.timings = []
-      start_section = 0
+      if self.album_art:
+        self.timings = [{"image_url":self.album_art,
+                          "transition_after": int(self.track.sections[0]['duration']*1000),
+                          "top_text": '',
+                          "bottom_text": '',
+                        }] 
+        start_section = 1
+      else:
+        self.timings = []
+        start_section = 0
+      if method == 'random_shuffle':
+        self.random_shuffle_memes(memes, start_section)
+      elif method == 'maximum_spread':
+        self.maximum_spread(memes, start_section)
+      else:
+        raise ValueError('unknown alignment method')
+  
+  def random_shuffle_memes(self, memes, start_section):
     shuffled_img = random.sample(memes.keys(), len(memes.keys()))
-    print shuffled_img
+    # print shuffled_img
     shuffled_phrases = random.sample(self.lyrics, len(self.lyrics))
     for section in self.track.sections[start_section:]:
       #draw cards
       key = shuffled_img.pop()
       img = memes[key]
-      print img
+      # print img
       img_path = img["source_image"]
       this_phrase = shuffled_phrases.pop()
       while len(this_phrase) < 2:
@@ -132,12 +153,58 @@ class memeMatcher:
                            "top_text": this_phrase[lyric_idx],
                            "bottom_text": this_phrase[lyric_idx+1],
                           })
-
-
+  
+  def maximum_spread(self, memes, start_section):
+    shuffled_img = random.sample(memes.keys(), len(memes.keys()))
+    sections_per_phrase = len(self.track.sections[start_section:-1])/len(self.lyrics)
+    current_section=start_section
+    for lyric_block in self.lyrics:
+      bars = []
+      for section in self.track.sections[current_section:current_section+sections_per_phrase]:
+        bars += [bar for bar in self.track.bars \
+                  if (bar[u'start']>=section[u'start']) and (bar[u'start']+bar[u'duration'])<=(section[u'start']+section[u'duration'])]
+      bars_per_linepair = len(bars)/(len(lyric_block)/2)
+      if bars_per_linepair <= 4:
+        bars_per_linepair = 4
+      else:
+        bars_per_linepair = 8
+      # bars_per_linepair = 4
+      for some_bars in grouper(bars_per_linepair,bars):
+        if some_bars.count(None)>0 or len(lyric_block)<2:
+          #combine with the last one and force last linepair of the phrase
+          self.timings[-1]["transition_after"] += int(sum([bar[u'duration'] for bar in bars if bar != None])*1000)
+          if len(lyric_block) > 1:
+            self.timings[-1]['bottom_text'] = lyric_block.pop(-1)
+            self.timings[-1]['top_text'] = lyric_block.pop(-1)
+        else:
+          key = shuffled_img.pop()
+          img = memes[key]
+          img_path = img["source_image"]
+          if len(shuffled_img)==0:
+            shuffled_img = random.sample(memes.keys(), len(memes.keys()))
+          top = lyric_block.pop(0)
+          bottom = lyric_block.pop(0)
+          self.timings.append({"image_url":img_path,
+                               "transition_after": int(sum([bar[u'duration'] for bar in some_bars])*1000),
+                               "top_text": top,
+                               "bottom_text": bottom,
+                              })
+    current_section += sections_per_phrase
+    #last section gets the album art again
+    self.timings.append({"image_url":self.album_art,
+                        "transition_after": int(self.track.sections[-1]['duration']*1000),
+                        "top_text": '',
+                        "bottom_text": '',
+                        })
+    
+    print self.track.sections
+    print self.track.bars
+    print self.timings
+    
 class memeMatcherTests(unittest.TestCase):
-	def setUp(self):
-		self.tester = memeMatcher('')
+  def setUp(self):
+    self.tester = memeMatcher('')
 
 
 if __name__ == '__main__':
-	unittest.main()
+  unittest.main()
